+++ toolkit/scaffold/project/cpp/qt5/cli.py
@@ -5,6 +5,7 @@ import sys
 import click
 
 from toolkit.scaffold.project.cpp.qt5.console import create_console
+from toolkit.scaffold.project.cpp.qt5.example import create_example
 
 
 @click.group(help="Create a cpp qt5 project scaffold.")
@@ -13,6 +14,8 @@ def create_cpp_qt5_project():
 
 
 create_cpp_qt5_project.add_command(create_console, "console")
+
+create_cpp_qt5_project.add_command(create_example, "example")
 
 if __name__ == "__main__":
     sys.exit(create_cpp_qt5_project())  # pragma: no cover